+++ addon/objects/uml-primitives/fd-uml-package.js
@@ -30,7 +30,9 @@ export default FdUmlElement.extend({
     @property attributes
     @type String
   */
-  attributes: Ember.computed.alias('primitive.Prop.Text'),
+  attributes: Ember.computed('primitive.Prop.Text', function() {
+    return this.get('primitive.Prop.Text').split('\n');
+  }),
 
   /**
     See {{#crossLink "FdUmlPrimitive/JointJS:method"}}here{{/crossLink}}.
@@ -92,6 +94,10 @@ export let Package = BaseClass.define('flexberry.uml.Package', {
           $buffer.css('font-weight', $input.css('font-weight'));
           $buffer.text($input.val());
           $input.width($buffer.width() + 1);
+          if (rect.type === 'header') {
+            newWidth = $input.width() / 0.8;
+          }
+
           if ($input.width() > newWidth) {
             newWidth = $input.width();
           }
@@ -121,7 +127,7 @@ export let Package = BaseClass.define('flexberry.uml.Package', {
 joint.shapes.flexberry.uml.PackageView = joint.shapes.flexberry.uml.BaseObjectView.extend({
   template: [
     '<div class="uml-class-inputs">',
-    '<input type="text" class="package-header-input header-input" value="" />',
+    '<textarea class="package-header-input header-input" value="" rows="1" wrap="off"></textarea>',
     '<textarea class="attributes-input body-input" value="" rows="1" wrap="off"></textarea>',
     '<div class="input-buffer"></div>',
     '</div>'
@@ -137,29 +143,23 @@ joint.shapes.flexberry.uml.PackageView = joint.shapes.flexberry.uml.BaseObjectVi
   initialize: function () {
     joint.shapes.flexberry.uml.BaseObjectView.prototype.initialize.apply(this, arguments);
     this.$box.find('.package-header-input').on('change', function (evt) {
-      this.model.set('name', Ember.$(evt.target).val());
-    }.bind(this));
-    this.$box.find('.package-header-input').on('input', function () {
-      this.model.updateRectangles();
+      let $textarea = Ember.$(evt.currentTarget);
+      let textareaText = $textarea.val();
+      let rows = textareaText.split(/[\n\r|\r|\n]/);
+      $textarea.prop('rows', rows.length);
+      this.model.set('name', textareaText);
     }.bind(this));
 
-    this.$box.find('.attributes-input').on('input', function (evt) {
+    this.$box.find('.package-header-input').on('input', function (evt) {
       let $textarea = Ember.$(evt.currentTarget);
       let textareaText = $textarea.val();
       let rows = textareaText.split(/[\n\r|\r|\n]/);
       $textarea.prop('rows', rows.length);
       this.model.updateRectangles();
-    }.bind(this));
-
-    this.$box.find('.attributes-input').on('change', function (evt) {
-      let $textarea = Ember.$(evt.currentTarget);
-      let textareaText = $textarea.val();
-      let rows = textareaText.split(/[\n\r|\r|\n]/);
-      $textarea.prop('rows', rows.length);
-      this.model.set('attributes', rows);
     }.bind(this));
 
     let upperInput = this.$box.find('.package-header-input');
+    upperInput.prop('rows', this.model.get('name').split(/[\n\r|\r|\n]/).length || 1);
     upperInput.val(this.model.get('name'));
   }
 });