+++ addon/objects/uml-primitives/fd-uml-template-class.js
@@ -48,10 +48,14 @@ export default FdUmlClass.extend({
 */
 export let TemplateClass = Class.define('flexberry.uml.TemplateClass', {
   attrs: {
+    rect: { 'width': 200 },
+
+    '.flexberry-uml-header-rect': { 'stroke': 'black', 'stroke-width': 1, 'fill': '#ffffff', 'fill-opacity': 0, 'mask': 'url(#custom-mask)' },
     '.flexberry-uml-params-rect': {
       'stroke': 'black', 'stroke-width': 1,
       'stroke-dasharray': '7 2',
-      'fill': 'white',
+      'fill': '#ffffff',
+      'fill-opacity': 0
     },
 
     '.flexberry-uml-params-text': {
@@ -65,16 +69,25 @@ export let TemplateClass = Class.define('flexberry.uml.TemplateClass', {
       'font-size': 12,
       'font-family': 'Arial'
     },
+
+    '.view-rect': { 'x': -1, 'y': -1, 'fill': 'white' },
+    '.not-view-rect': { 'x': -1, 'y': -1, 'fill': 'black' }
   },
 
   params: [],
 }, {
   markup: [
     '<g class="rotatable">',
-    '<g class="scalable">',
-    '<rect class="flexberry-uml-header-rect"/><rect class="flexberry-uml-params-rect"/>',
-    '<rect class="flexberry-uml-body-rect"/><rect class="flexberry-uml-footer-rect"/>',
-    '</g>',
+    '<defs>',
+    '<mask id="custom-mask">',
+    '<rect class="view-rect"/>',
+    '<rect class="not-view-rect"/>',
+    '</mask>',
+    '</defs>',
+    '<rect class="flexberry-uml-header-rect"/>',
+    '<rect class="flexberry-uml-params-rect"/>',
+    '<rect class="flexberry-uml-body-rect"/>',
+    '<rect class="flexberry-uml-footer-rect"/>',
     '</g>'
   ].join(''),
 
@@ -87,15 +100,68 @@ export let TemplateClass = Class.define('flexberry.uml.TemplateClass', {
       { type: 'body', text: this.get('attributes'), element: this },
       { type: 'footer', text: this.get('methods'), element: this }
     ];
+  },
+
+  updateRectangles() {
+    let rects = this.getRectangles();
+    let offsetY = 0;
+    let newHeight = 0;
+    let newWidth = 0;
+    let paramHeight = 0;
+    rects.forEach(function(rect) {
+      if (this.markup.includes('flexberry-uml-' + rect.type + '-rect') && rect.element.inputElements) {
+        let $buffer = rect.element.inputElements.find('.input-buffer');
+        let rectHeight = 0;
+        let inputs = rect.element.inputElements.find('.' + rect.type + '-input');
+        inputs.each(function() {
+          let $input = Ember.$(this);
+          $buffer.css('font-weight', $input.css('font-weight'));
+          $buffer.text($input.val());
+          $input.width($buffer.width() + 1);
+          if (rect.type === 'params') {
+            paramHeight = $input.height() + 4;
+            rect.element.attr('.flexberry-uml-params-rect/width', $input.width() + 20);
+            rect.element.attr('.not-view-rect/width', $input.width() + 20);
+            rect.element.attr('.not-view-rect/height', paramHeight);
+          } else if ($input.width() > newWidth) {
+            newWidth = $input.width();
+          }
+
+          rectHeight += $input.height();
+        });
+
+        rectHeight += rect.element.get('heightBottomPadding') || 0;
+        newHeight += rectHeight;
+        rect.element.attr('.flexberry-uml-' + rect.type + '-rect/height', rectHeight);
+        rect.element.attr('.flexberry-uml-' + rect.type + '-rect/transform', 'translate(0,' + offsetY + ')');
+
+        offsetY += rectHeight;
+      }
+    }, this);
+
+    newWidth += (this.get('widthPadding') || 0) * 2;
+    rects.forEach(function(rect) {
+      if (rect.type === 'params') {
+        rect.element.attr('.flexberry-uml-params-rect/transform', 'translate(' + (newWidth - 10) + ',15)');
+        rect.element.attr('.not-view-rect/transform', 'translate(' + (newWidth - 10) + ',' + (15 - paramHeight) + ')');
+      } else {
+        rect.element.attr('.flexberry-uml-' + rect.type + '-rect/width', newWidth);
+      }
+    });
+
+    this.attr('.view-rect/width', newWidth + 2);
+    this.attr('.view-rect/height', newHeight + 2);
+
+    this.resize(newWidth, newHeight);
   }
 });
 
 joint.shapes.flexberry.uml.TemplateClassView = joint.shapes.flexberry.uml.ClassView.extend({
   template: [
     '<div class="uml-class-inputs">',
-    '<textarea type="text" class="params-input" value="" rows="1" wrap="off"> </textarea>',
-    '<input type="text" class="class-name-input header-input class-t" value="" />',
-    '<input type="text" class="class-stereotype-input header-input" value="" />',
+    '<textarea class="params-input" value="" rows="1" wrap="off"> </textarea>',
+    '<textarea class="class-name-input header-input" value="" rows="1" wrap="off"></textarea>',
+    '<textarea class="class-stereotype-input header-input" value="" rows="1" wrap="off"></textarea>',
     '<textarea class="attributes-input body-input" value="" rows="1" wrap="off"></textarea>',
     '<textarea class="methods-input footer-input" value="" rows="1" wrap="off"></textarea>',
     '<div class="input-buffer"></div>',
@@ -104,15 +170,57 @@ joint.shapes.flexberry.uml.TemplateClassView = joint.shapes.flexberry.uml.ClassV
 
   initialize: function() {
     joint.shapes.flexberry.uml.ClassView.prototype.initialize.apply(this, arguments);
-    this.$box.find('.params-input').on('input', function () {
+    this.$box.find('.params-input').on('input', function (evt) {
+      let $textarea = Ember.$(evt.currentTarget);
+      let textareaText = $textarea.val();
+      let rows = textareaText.split(/[\n\r|\r|\n]/);
+      $textarea.prop('rows', rows.length);
       this.model.updateRectangles();
     }.bind(this));
 
     this.$box.find('.params-input').on('change', function (evt) {
-      this.model.set('params', Ember.$(evt.target).val());
+      let $textarea = Ember.$(evt.currentTarget);
+      let textareaText = $textarea.val();
+      let rows = textareaText.split(/[\n\r|\r|\n]/);
+      $textarea.prop('rows', rows.length);
+      this.model.set('params', textareaText);
     }.bind(this));
 
     let paramsInput = this.$box.find('.params-input');
+    paramsInput.prop('rows', this.model.get('params').split(/[\n\r|\r|\n]/).length || 1);
     paramsInput.val(this.model.get('params'));
-  }
+  },
+
+  updateBox: function() {
+    // Set the position and dimension of the box so that it covers the JointJS element.
+    let bbox = this.model.getBBox();
+    let paramsBox = this.$box.find('.params-input');
+    this.$box.css({
+      width: bbox.width,
+      height: bbox.height,
+      left: bbox.x,
+      top: bbox.y + paramsBox.height() + 4,
+      transform: 'rotate(' + (this.model.get('angle') || 0) + 'deg)'
+    });
+
+    paramsBox.css({
+      left: bbox.width,
+      top: 15 - paramsBox.height(),
+      position: 'absolute'
+    });
+  },
+
+  render: function() {
+    joint.shapes.flexberry.uml.BaseObjectView.prototype.render.apply(this, arguments);
+
+    let mask = document.getElementById('custom-mask');
+    let viewMaskId = Ember.$(mask).children('.view-rect').attr('id');
+    let maskId = 'mask_tc_' + viewMaskId;
+    mask.setAttribute('id', maskId);
+    let attrs = this.model.get('attrs');
+    attrs['.flexberry-uml-header-rect'].mask = 'url(#' + maskId + ')';
+    this.model.updateRectangles();
+
+    return this;
+  },
 });